% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_columns.R
\name{check_columns}
\alias{check_columns}
\title{Given a set of columns of data in a time series check whether they are all
of the same mode, class, and if factor or string-valued, the same collection of
strings.}
\usage{
check_columns(data_, col_names)
}
\arguments{
\item{data_}{The dataframe to check.}

\item{col_names}{Character vector of columns to check.}
}
\value{
a ggplot2 object for the columns that were checked
}
\description{
It also emits the output of skimr::skim and a plot so that the user can also
visually inspect additional aspects of the columns.
}
\details{
Note that the returned plot can be captured and modified further to, for example,
change the X and Y limits.
}
\examples{
data_that_passes <- data.frame(smk80 = c('never', 'past', 'never'),
                               smk82 = c('past', 'never', 'past'),
                               smk84 = c('never', 'never', 'past'))
data_that_fails <- data.frame(smk80 = c('never smoker', 'past smoker', 'never smoker'),
                              smk82 = c('past', 'never', 'past'),
                              smk84 = c('never', 'never', 'past'))
check_columns(data_that_passes, generate_column_names('smk', from = 1980, to = 1984))
\dontrun{

check_columns(data_that_fails, generate_column_names('smk', from = 1980, to = 1984))

}

# You can also capture the plot returned and modify it.
numeric_data_to_plot <- data.frame(
                           bmi80 = rnorm(200, mean = 30, sd = 3),
                           bmi82 = rnorm(200, mean = 31, sd = 2.8),
                           bmi84 = rnorm(200, mean = 32, sd = 2.5))
p <- check_columns(numeric_data_to_plot,
                   generate_column_names('bmi', from = 1980, to = 1984))
# Show default plot
p

# Modify default plot to zoom in by changing the limits
library(ggplot2)
p + ylim(27, 35)

# Modify the default plot to invert the X and Y axes and use
# a different theme.
p + coord_flip() + theme_minimal()

}
