% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_column_names.R
\name{generate_column_names}
\alias{generate_column_names}
\title{Generate the column names from a prefix for our time periods of interest.}
\usage{
generate_column_names(prefix, from = 1976, to = 2018)
}
\arguments{
\item{prefix}{The prefix of the variable name.}

\item{from}{The starting year, as an integer, for our time periods of interest.}

\item{to}{The ending year, as an integer, for our time periods of interest.}
}
\value{
A character vector of the column names.
}
\description{
Generate the column names from a prefix for our time periods of interest.
}
\examples{
generate_column_names(prefix = 'irt', from = 1990, to = 2002)
}
