% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_markdown.R
\name{print_table1}
\alias{print_table1}
\title{Convenience method to output a summary_table object as markdown.}
\usage{
print_table1(summary_table, title = "**Table 1**", caption = "",
  output_file)
}
\arguments{
\item{summary_table}{The qwraps2::summary_table object to print.}

\item{title}{The title for the table.}

\item{caption}{The caption for the table.}

\item{output_file}{(optional) The file to which to write the markdown.}
}
\description{
Markdown is emitted to the console, and optionally to an external file.
}
\details{
This function takes as input a \code{\link[qwraps2:summary_table]{qwraps2::summary_table()}} object`. See the
documentation and vignettes of that package to learn how to use qwraps2.
}
\examples{
require(qwraps2)

mtcars_summaries <-
list('Miles Per Gallon' =
       list('min' = ~ min(mpg),
       'mean' = ~ mean(mpg),
       'mean &plusmn; sd' = ~ qwraps2::mean_sd(mpg),
       'max' = ~ max(mpg)),
     'Weight' =
       list('median' = ~ median(wt)),
     'Cylinders' =
       list('4 cyl: n (\%)' = ~ qwraps2::n_perc0(cyl == 4),
            '6 cyl: n (\%)' = ~ qwraps2::n_perc0(cyl == 6),
            '8 cyl: n (\%)' = ~ qwraps2::n_perc0(cyl == 8)))

print_table1(summary_table = qwraps2::summary_table(mtcars, mtcars_summaries),
             title = '**mtcars summary table**',
             caption = 'example using mtcars')
}
\seealso{
\code{\link[qwraps2:summary_table]{qwraps2::summary_table()}}, \code{browseVignettes('qwraps2')}
}
